requirements.txt
flask
threading
json
time
azure.iot.hub
azure.eventhub